EFS-698 Minor fixes
