+++ Makefile
@@ -12,6 +12,7 @@ publish:
 
 .PHONY:up
 up:
+	yarn build_react && \
 	docker compose -f docker-compose.yml -p fhir-dev-ui build --parallel && \
 	docker compose -p fhir-dev-ui -f docker-compose.yml up --detach && \
 	echo FHIR server: http://localhost:5051/admin
