+++ src/src/App.tsx
@@ -1,8 +1,8 @@
 import './App.css';
 import React, { useEffect, useState } from 'react';
 import { Routes, Route, createBrowserRouter, RouterProvider } from 'react-router-dom';
-import HomePage from './HomePage';
-import AboutPage from './AboutPage';
+import HomePage from './pages/HomePage';
+import AboutPage from './pages/AboutPage';
 
 import PatientChatGptPage from './PatientChatGptPage';
 import ObservationGraph from './ObservationGraph';
@@ -15,13 +15,14 @@ import PersonPatientLinkPage from './admin/personPatientLink';
 import SearchLogsPage from './admin/searchLogs';
 import SearchPage from './pages/SearchPage';
 import IndexPage from './pages/IndexPage';
-import ErrorPage from './ErrorPage';
+import ErrorPage from './pages/ErrorPage';
 import EnvironmentContext from './EnvironmentContext';
 import axios, { AxiosResponse } from 'axios';
 import { jwtParser } from './utils/jwtParser';
 import { TUserDetails } from './types/baseTypes';
 import Auth from './pages/Auth';
 import { getCookie } from './utils/cookie.utils';
+import Logout from './pages/logout';
 
 type TApiEnvResponse = {
     FHIR_SERVER_URL: string;
@@ -44,6 +45,7 @@ function App(): React.ReactElement {
         return (
             <Routes>
                 <Route path="/" element={<HomePage />} />
+                <Route path="/logout" element={<Logout />} />
                 {isLoggedIn || (env && !env.AUTH_ENABLED) ? (
                     <>
                         <Route path="/about" element={<AboutPage />} />
@@ -53,12 +55,18 @@ function App(): React.ReactElement {
                         <Route path="/observationGraph" element={<ObservationGraph />} />
                         <Route path="/observationTimeline" element={<ObservationTimeline />} />
                         <Route path="/4_0_0/:resourceType/_search/*" element={<SearchPage />} />
-                        <Route path="/4_0_0/:resourceType/:id?/:operation?/*" element={<IndexPage />} />
+                        <Route
+                            path="/4_0_0/:resourceType/:id?/:operation?/*"
+                            element={<IndexPage />}
+                        />
                         <Route path="/4_0_0/:resourceType/:operation?/*" element={<IndexPage />} />
                         <Route path="/admin" element={<AdminIndexPage />} />
                         <Route path="/admin/personMatch/*" element={<PersonMatchPage />} />
                         <Route path="/admin/patientData/*" element={<PatientDataPage />} />
-                        <Route path="/admin/personPatientLink/*" element={<PersonPatientLinkPage />} />
+                        <Route
+                            path="/admin/personPatientLink/*"
+                            element={<PersonPatientLinkPage />}
+                        />
                         <Route path="/admin/searchLog/*" element={<SearchLogsPage />} />
                     </>
                 ) : (
@@ -82,7 +90,8 @@ function App(): React.ReactElement {
                     // noinspection ExceptionCaughtLocallyJS
                     throw new Error('Network response was not ok');
                 }
-                const { FHIR_SERVER_URL, AUTH_CUSTOM_GROUP, AUTH_CUSTOM_SCOPE, ...envs } = response.data;
+                const { FHIR_SERVER_URL, AUTH_CUSTOM_GROUP, AUTH_CUSTOM_SCOPE, ...envs } =
+                    response.data;
                 setFhirUrl(FHIR_SERVER_URL);
                 setEnv(envs);
                 const { username, scope, isAdmin } = jwtParser({
